Refactor postgres test base
